Match path of config in bidsapp mode to parent
The config file in bidsapp mode was stored in the `code` directory in
the output folder. This introduced a breaking change, as `Snakefiles`
would need to refer to the config file via `workflow.source_path()`.
This adjusts the path to match that of the parent app:
    `parent/config/snakebids.yml` -> `output/config/snakebids.yml`

This way, 0.4.0 will not be breaking.

See #66