use action to check PR label
Uses `pr-label-checker` to check for assigned PR labels - specifically whether it contains the `skip_changelog` label. Previous implementation would only check label when the PR label was assigned, deleted, or editted, but would not persist.

If the PR does not contain the `skip_changelog` label, continue with rest of the workflow.