Make _generate_search_terms more generic
_generate_search_terms is currently a boutique function for finding
Pybids filters for subjects (i.e. it returns the output in a dict with a
"subjects" key). This makes the function more generic, now generating
Pybids filters for any arbitrary values. It returns a list of Pybids
filter strings and a boolean indicating whether regex_search must be
enabled in Pybids (which is the case whne exclude values are provided).

The function name is changed to _generate_filters