Refactor process_layout_wildcard
The function is renamed to _parse_bids_path to be consistent with
_parse_custom_path.

Rather than processing a single wildcard at a time, the function now
processes the entire list of wildcards. This will help relieve some of
the deep nesting in get_lists_from_bids, especially when this function
is refactored later.