Implement BidsLists
BidsLists is a NamedTuple that will serve as an intermediary between
Path parsing and BidsInputs. It stores each input descriptor along with
its input_path, input_zip_lists, input_lists, and input_wildcards.

A classmethod is implemented on BidsInputs that can read an iterable of
these BidsLists and construct a BidsInputs. This method relies on a new
@staticmethod that converts BidsLists into dicts. This second method has
an associated test.