Allow generate_inputs to return BidsInputs
generate_inputs can now return BidsInputs if the new use_bids_inputs parameter
is set to True. Be default, it will return the old dict and log a
warning that the dict is being deprecated and the default will switch in
a future update. The parameter can be set to False to ensure the dict is
always returned and to suppress the warning.

Internally, generate_inputs already uses BidsInputs. It returns a Dict
by calling the as_dict property on BidsInputs.

_get_lists_from_bids has been refactored as a Generator of BidsLists.
These BidsLists can be directly consumed by BidsInputs to construct an
instance.

A copy of test_t1w has been created and modified to test the new
BidsInputs syntax. test_get_lists_from_bids has been modified to account
for the new return of BidsLists

Also expands documentation of generate_inputs, explaining in more detail
what the filters Dict in pybids_inputs should contain

Expand documentation of generate_inputs()

Resolves #100