Optimize filter_list for speed
The previous edit to filter_list lowered the speed by half. This
restores the function much to it's original form, using sets now for
improved readability.

zip_lists are once again returned as lists rather than tuples

A previous bug in the original filter caused incorrect results in some
cases. If multiple keys were supplied, and the first key did not produce
any matches, but the second key did, the matches from the second key
would be incorrectly returned.

New to this version: if none of the filters provided are keys in the
zip_list, the entire zip_list will be returned