Remove explicit workflow mode
Remove any public-facing references to workflow/bidsapp mode.

When the output dir is set to `snakebids_root` or any subdirectory equal
or under to `snakebids_root/results`, the cwd will remain in the
snakebids_root (the config and .snakemake folders will be saved there).

When the output is set to `snakebids_root`, it will automatically be
switched to `snakebids_root/results`

Any other path will operate exactly as "bidsapp" mode use to work.

- Add new CLI flag --force-output to force writing results into a new
  directory with contents (e.g. a directory without a .snakebids folder)
- Such a force is no longer necessary for any folders under
  `snakebids_root/results`
- Add deprecation warnings to  --workflow-mode, --retrofit, and
  --force-conversion. Remove their entries from SnakebidsArgs. They will
  no longer appear in CLI help, but will print a warning if provided.
  Change SnakebidsArgs to receive arguments from the new --force-output
- All unused functions and tests are deleted
- Combine all app tests into a single, hypothesis driven test