Add minimal fixes to get logging working
There was previously no central initializing of handlers for logging, so
only WARNING and higher was getting logged and without a modicum of
formatting.

Add a centralized _logging module to initialize the base logger. This
gets called at the two app entrypoints: snakebids.admin:main and upon
initialization of SnakebidsApp