Block components with same names in dataset init
BidsDataset.from_iterable was blindly converting the iterable of
components passed into a dict. If any of these components had the same
name, they would overlap in the dict (since the name is used as a key),
and the first would be silently removed.

Fix this by checking for duplicate names.

Resolves #181