Use entity-specific patterns in _parse_bids_path
This corrects some of the mistakes made by the previous parser: mistakes
caused by overlapping entity names (space and ce), inability to have
"datatype" or "extension" as entities, and paths with just suffixes.

Now use the expanded bids_tags.json file to specifically match every
tag. If a tag is given that isn't found in bids_tags.json, fall back to
the previous `tag-[a-zA-Z0-1]` syntax.

Resolves #192