Implement pretty printing for dataset objects
BidsComponents and BidsDatasets previously used the default formatting
provded by attrs and dict respectively, which was complete but
unreadable. This commit adds newlines and aligns the zip_list values
into a table. If in an interactive display, zip_list lists will be
automatically elided to fit within the display.

Adds full test suite for new functionality

Resolves #197