Update cookiecutter template
This commit updates the cookiecutter variables, restructures cookiecutter prompts, and makes creation of docs optional (while moving .rst files to .md). Additionally, cookiecutter post_gen_hooks are used to update the the snakebids version required to the current installed snakebids version via `importlib`.

Additionally, this commit introduces changes in the unmerged draft PR #171, moving the
workflow files into the package directory and changing the `run.py` to
executable. As noted in that PR, this fixes a bug in resolving path in
`run.py`, as well as allows for proper inclusion of workflow files in
python packaging. The associated test has also been updated to reflect
this (as well as cookiecutter related changes).