Fix pybidsdb path resolution
The resolved path was getting overwritten by the args_dict in
SnakebidsArgs. This didn't happen previously because the resolved path
was saved under a different name than the CLI parameter, but since
resolving the names, the CLI parameter name overwrites the resolved
name.

Fix by changing the order of update_config and the manual setting of the
pybids path