Update dependencies and pyright version
Fix a number of type issues arising with latest pyright

Need to lock pyright to >=1.1.324 because previous versions have a false
positive bug

Some dict.get(..., {}) require laborious workarounds because the value
type of the dict is a TypedDict, which is different than a regular dict
and can't be constructed easily on the fly. The latest pyright version
uses the latest typeshed, which changed the inference rules to prevent
our previous "hack" from working. This situation will probably never
change.

Broadened the annotated types for some keyword args on functions
accepting both fixed and generic kwargs with differing types. The fixed
kwargs need to allow at least the same types as the generic args, or we
get type issues downstream. Use type validation to compensate