Fix formatting of empty zip lists
Empty zip lists (meaning zip lists with entities but without any
values) gave an IndexError when attempting formatting. This fixes the
problem, plus refactors the formatting code. In particular, the function
to find the elision has been refactored to work only on the 'values'
portion of the table to reduce the amount of 'magic numbers', previously
needed to account for the table keys

Resolves #339