Create documentation compatibility for old apps
Include add_dynamic_args as deprecated placeholder. Some apps were using
it to prepare a documentation cli. Include it with a deprecation warning
for a more graceful change. It's otherwise a no-op, it's funcationality
subsumed by SnakeBidsApp and the underlying bidsapp

Expose .parser and .config on SnakeBidsApp

Lazily construct app in SnakeBidsApp: Improves compatibility for some
tests