Update argument handling for --derivatives
The initial implementation of bidsargs had --derivatives take the same
arguments as previously specified in the template app: action="+". This
made the argument alway look for one or more paths. This does not
reflect the pybids api, which additionally takes True and False

The fix adjusts this so that --derivatives defaults to False, if
provided by itself give True, and if provide with paths, gives a list of
Path().