Prevent _run-{values} with leading 0s
When used with `run`, the leading 0 gets stripped because run is
converted into an int. This breaks tests that instantiate a datset on
the filesystem and re-parse it using generate_inputs.

Adds a `type` field to run in `bids_tags.json` and uses this to filter
integers with leading 0s in the zip_lists strategy