Transition from zip-list to entry core
The entry table uses entry-first strides. This will be a much more
efficient storage method for future API that makes individual entries
more accessible. As much of the core as is currently feasible is
rewritten to use entries over zip-lists

The public API is completely unchanged as of now.

Tests for filtering are largely unchanged, although some may be
redundant at this point.